+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased](https://github.com/schemathesis/schemathesis/compare/v4.21.0...HEAD) - TBD
 
+### :bug: Fixed
+
+- Missing boundary negative in the coverage phase for boolean `exclusiveMinimum` / `exclusiveMaximum`.
+
 ## [4.21.0](https://github.com/schemathesis/schemathesis/compare/v4.20.3...v4.21.0) - 2026-06-01
 
 ### :rocket: Added